acis_insurance_analytics copy/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
xgboost>=2.0.0
shap>=0.44.0
pyyaml>=6.0.0
dvc>=3.0.0
dvc[ssh]
